Add/rtenvsub toggle switch (#69)
Add toggle switch for dev mode

+++ docker_utils/rtenvsub.sh
@@ -20,6 +20,7 @@ version='0.1'
 # Set a safe umask
 umask 0077
 
+dev_mode="${DEV_MODE:-false}"
 default_library_name='shtdlib.sh'
 default_base_download_url='https://raw.githubusercontent.com/sdelements/shtdlib/master'
 default_install_path='/usr/local/bin'
@@ -104,7 +105,7 @@ if ! whichs envsubst ; then
     color_echo cyan 'Perhaps this can be fixed with: apt-get -y install gettext-base'
     exit 1
 fi
-if ! whichs inotifywait ; then
+if ! whichs inotifywait && ${dev_mode} ; then
     color_echo red "Unable to locate the inotifywait command, please make sure it's available"
     color_echo cyan 'Perhaps this can be fixed with: apt-get install inotify-tools'
     exit 1
@@ -398,7 +399,9 @@ function mirror_envsubst_paths {
             color_echo magenta "Destination directory does not contain any files, no pipes created for ${full_path}!"
         else
             for file in "${files[@]:-}"; do
-                add_on_sig "rm -f ${destination}${file#${full_path}}"
+                if ${dev_mode} ; then
+                    add_on_sig "rm -f ${destination}${file#${full_path}}"
+                fi
                 if ${nofifo} ; then
                     render_file "${destination}" "${file}" "${full_path}"
                 else
@@ -407,20 +410,22 @@ function mirror_envsubst_paths {
             done
         fi
 
-        # Set up safe cleanup for directory structure (needs to be done in
-        # reverse order to ensure safety of operation without recursive rm
-        local index
-        for (( index=${#directories[@]}-1 ; index>=0 ; index-- )) ; do
-            add_on_sig "rmdir ${destination}${directories[${index}]#${full_path}}"
-        done
+        if ${dev_mode} ; then
+            # Set up safe cleanup for directory structure (needs to be done in
+            # reverse order to ensure safety of operation without recursive rm
+            local index
+            for (( index=${#directories[@]}-1 ; index>=0 ; index-- )) ; do
+                add_on_sig "rmdir ${destination}${directories[${index}]#${full_path}}"
+            done
 
-        # Run update loop and detach it
-        if ${daemonize} ; then
-            inotify_looper "${destination}" "${full_path}" &
-        else
-            inotify_looper "${destination}" "${full_path}" &
+            # Run update loop and detach it
+            if ${daemonize} ; then
+                inotify_looper "${destination}" "${full_path}" &
+            else
+                inotify_looper "${destination}" "${full_path}" &
+            fi
+            looper_pids+=( "${!}" )
         fi
-        looper_pids+=( "${!}" )
     done
     if ! ${daemonize} ; then
         debug 8 "Waiting for looper pids: ${looper_pids[*]}"
